tests/wpt/web-platform-tests/old-tests/submission/Opera/script_scheduling/043.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: DOM mutation events when adding external scripts: DOMNodeInserted </title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>log('inline script #1');
		document.addEventListener( 'DOMNodeInserted', function(){ log('DOMNodeInserted'); }, false );
		testlib.addScript('', { src: 'scripts/include-1.js' }, document.getElementsByTagName('head')[0], false);
		log('end script #1');
	</script>

	<script type="text/javascript">
	log( 'inline script #2' );
	var t = async_test()

	function test() {
		assert_any(assert_array_equals, eventOrder, [['inline script #1', 'DOMNodeInserted', 'end script #1', 'external script #1', 'inline script #2'],
                                                             ['inline script #1', 'DOMNodeInserted', 'end script #1', 'inline script #2', 'external script #1']]
                          );
		t.done();
        }
	onload = t.step_func(test)
	</script>

</body></html>
